Add the --non-interactive flag when installing type stubs via mypy
The --non-interactive flag forces mypy to hide the errors about missing stubs and instead simply install stubs without asking for confirmation.  It also does not return an error code, which it does without this flag even if you opt to let it install the missing type stubs.

Also, move mypy command toward the end of the script, which allows all the other good things to happen even if the mypy
command fails.

Co-authored-by: Jeremy Frasier <jeremy.frasier@gwe.cisa.dhs.gov>
Co-authored-by: Mark Feldhousen <mark.feldhousen@gwe.cisa.dhs.gov>