Remove an unnecessary "Optional" type hint
This was causing mypy 1.13 to throw an error like this:

src/cyhy_cvesync/main.py:63: error: Argument 3 to "process_urls" has incompatible type "int | None"; expected "int"  [arg-type]

The current version of mypy used in by pre-commit is 1.10.0 and it was NOT flagging this error, but my local mypy 1.13.0 was so I went ahead and fixed it.  All pytests still pass after this change.